Move ValueTable and WitnessM4 into binius-core
Both are witness data over a `ValueVec`, which is core's, but both lived in
binius-m4-prover because the entry points that build them need a circuit to
evaluate. Core cannot depend on the frontend — the frontend depends on core —
so the types move down and the constructors invert into methods on the
frontend's own types.

`ValueTable` keeps its storage, accessors and `instance_value_vec` in core,
and gains `from_hidden_words`: the seam the frontend fills, taking the
wire-major hidden buffer whose row count the layout already implies.
Population becomes `Circuit::populate_batch`, `populate_batch_parallel` and
`populate_batch_parallel_with_stripe_width`, with `BatchWitnessFiller`
alongside them. `ValueTable::populate(&circuit, k, fill)` reads
`circuit.populate_batch(k, fill)` now, which puts the circuit where a reader
looks for it.

`WitnessM4` keeps its fields and `verify` in core, next to the
`ConstraintSystemM4` that checks it and the `ChipInstances` accessor it is
served through. Generation becomes `CircuitM4::generate_witness`, carrying
`PopulateM4Error` with it since that error wraps the frontend's own.

Two operations stay in binius-m4-prover as free functions over a borrowed
table: `commit_layout` and `pack_table` reach binius-prover,
binius-m4-verifier and binius-compute, none of which core can depend on, and
neither has anything to do with building a table.

The tests follow their subjects. The batch-population tests move to the
frontend, as does the chip-accelerated CRC-64 integration test. The CRC-64
population test stays with the prover, since its fixture is that crate's
`test_utils`, which `prove`, `shift` and `instance_fold` still use.

The two new inherent-impl blocks carry `allow(clippy::multiple_inherent_impl)`,
following `ops.rs`, which keeps the builder's word operations out of the
compiler module the same way.
