+++ crates/core/src/constraint_system/mod.rs
@@ -9,6 +9,7 @@ mod proof;
 mod shift;
 mod system;
 mod value_index;
+mod value_table;
 mod value_vec;
 mod values_data;
 
@@ -18,5 +19,6 @@ pub use proof::*;
 pub use shift::*;
 pub use system::*;
 pub use value_index::*;
+pub use value_table::*;
 pub use value_vec::*;
 pub use values_data::*;
