+++ crates/m4-prover/benches/keccak_witness_gen.rs
@@ -10,7 +10,6 @@ use std::array;
 use binius_circuits::keccak::permutation::keccak_f1600;
 use binius_core::word::Word;
 use binius_frontend::{Circuit, CircuitBuilder, Wire};
-use binius_m4_prover::ValueTable;
 use criterion::{Criterion, criterion_group, criterion_main};
 
 /// The base-2 logarithm of the instance count: 2^13 = 8192 instances.
@@ -57,29 +56,30 @@ fn bench_keccak_witness_gen(c: &mut Criterion) {
 
 	group.bench_function("value_table", |b| {
 		b.iter(|| {
-			ValueTable::populate(&circuit, LOG_INSTANCES, |instance, w| {
-				for lane in 0..STATE_LANES {
-					w[input[lane]] = input_word(instance, lane);
-				}
-			})
-			.unwrap()
+			circuit
+				.populate_batch(LOG_INSTANCES, |instance, w| {
+					for lane in 0..STATE_LANES {
+						w[input[lane]] = input_word(instance, lane);
+					}
+				})
+				.unwrap()
 		});
 	});
 
 	for stripe_width in STRIPE_WIDTHS {
 		group.bench_function(format!("value_table_parallel_{stripe_width}"), |b| {
 			b.iter(|| {
-				ValueTable::populate_parallel_with_stripe_width(
-					&circuit,
-					LOG_INSTANCES,
-					stripe_width,
-					|instance, w| {
-						for lane in 0..STATE_LANES {
-							w[input[lane]] = input_word(instance, lane);
-						}
-					},
-				)
-				.unwrap()
+				circuit
+					.populate_batch_parallel_with_stripe_width(
+						LOG_INSTANCES,
+						stripe_width,
+						|instance, w| {
+							for lane in 0..STATE_LANES {
+								w[input[lane]] = input_word(instance, lane);
+							}
+						},
+					)
+					.unwrap()
 			});
 		});
 	}
