+++ crates/m4-prover/src/lib.rs
@@ -1,9 +1,12 @@
 // Copyright 2025 Irreducible Inc.
 // Copyright 2026 The Binius Developers
 
-//! Witness table and prover for the data-parallel Binius64 M4 proof system.
+//! Prover for the data-parallel Binius64 M4 proof system.
+//!
+//! The witness it proves over is a [`ValueTable`](binius_core::ValueTable), which
+//! [`binius_core`] defines and [`binius_frontend`] populates. This crate commits one and reduces
+//! the constraint system over it.
 
-mod m4_witness;
 mod prove;
 mod shift;
 #[cfg(test)]
@@ -11,7 +14,6 @@ mod test_utils;
 mod value_table;
 mod witness;
 
-pub use m4_witness::{PopulateM4Error, WitnessM4};
 pub use prove::{IOPProver, Prover};
-pub use value_table::{BatchWitnessFiller, ValueTable};
+pub use value_table::{commit_layout, pack_table};
 pub use witness::OperandColumns;
